# fix: TypeScript performance of `SelectionCard.tsx`