--- a/packages/manager/.changeset/pr-11077-tech-stories-1728500762562.md
+++ b/packages/manager/.changeset/pr-11077-tech-stories-1728500762562.md
@@ -1,0 +1,5 @@
+---
+"@linode/manager": Tech Stories
+---
+
+Fix TypeScript performance of `SelectionCard.tsx` ([#11077](https://github.com/linode/manager/pull/11077))