## Description 📝

- Fixes typescript performance of `SelectionCard.tsx` 🏎️ 

## Preview 📷

| Before  | After   |
| ------- | ------- |
| ![Screenshot 2024-10-09 at 3 03 54 PM](https://github.com/user-attachments/assets/c71d2f5f-b304-4187-b060-c08b73c1c07e) | ![Screenshot 2024-10-09 at 3 00 09 PM](https://github.com/user-attachments/assets/785d1456-d8a1-4259-b7c2-d5b1690eb149) |
| type-checking takes 5 seconds |  type-checking takes 13ms | 

## How to test 🧪

- Verify type checking passes in Github Actions
- Verify there are no regressions with the component, it's types, and it's behavior
- Profile this if you want to (`yarn tsc --generateTrace ./trace --incremental false` and inspect with `chrome://tracing`)

## As an Author I have considered 🤔

- [ ] 👀 Doing a self review
- [ ] ❔ Our [contribution guidelines](https://github.com/linode/manager/blob/develop/docs/CONTRIBUTING.md)
- [ ] 🤏 Splitting feature into small PRs
- [x] ➕ Adding a [changeset](https://github.com/linode/manager/blob/develop/docs/CONTRIBUTING.md#writing-a-changeset)
- [ ] 🧪 Providing/Improving test coverage
- [ ] 🔐 Removing all sensitive information from the code and PR description
- [ ] 🚩 Using a feature flag to protect the release
- [ ] 👣 Providing comprehensive reproduction steps
- [ ] 📑 Providing or updating our documentation
- [ ] 🕛 Scheduling a pair reviewing session
- [ ] 📱 Providing mobile support
- [ ] ♿  Providing accessibility support